added glossary and help button and pop up

+++ client/src/utils/helpers.js
@@ -1,10 +1,11 @@
 import axios from "axios";
+// const pyBackEnd = typeof process == 'object' ? process.env.BACK : "http://0.0.0.0:8000";
 import { SET_STOCK_WEIGHTS } from "./actions";
 import Auth from "../utils/auth";
-
 const pyBackEnd = "https://pern-portfolio-backend-805cd64a428d.herokuapp.com";
 
-// Index options object
+// const pyBackEnd = "http://0.0.0.0:8000"
+
 export const indexOptions = {
   "SP500": "^GSPC",
   "Dow Jones": "^DJI",
@@ -32,16 +33,19 @@ console.log(indexOptions["SP500"]);
 
 export async function stockData(stockSymbol, startDate, endDate) {
   console.log("stockSymbol", stockSymbol);
-  const response = await axios.get(`${pyBackEnd}/stockgraph`, {
-    params: {
-      symbol: stockSymbol,
-      start: startDate,
-      end: endDate,
-    },
-  });
+  const response = await axios.get(
+    `${pyBackEnd}/stockgraph`,
+
+    {
+      params: {
+        symbol: stockSymbol,
+        start: startDate,
+        end: endDate,
+      },
+    }
+  );
   return response.data;
 }
-
 export async function stockInfo(stockSymbol) {
   const response = await axios.get(`${pyBackEnd}/stockinfo`, {
     params: {
@@ -60,6 +64,7 @@ export async function linReg(stockSymbol, searchIndex, startDate, endDate, weigh
       start: startDate,
       end: endDate,
       stockWeights: weights,
+
     },
   });
   return response.data;
@@ -87,62 +92,14 @@ export async function getStockWeightsIdb() {
 }
 
 export async function getCompanyFinancials(stockSymbol, statement, quarterly) {
-  const url = `${pyBackEnd}/financials`;
-  try {
-    const response = await axios.get(url, {
-      params: {
-        symbol: stockSymbol,
-        statement: statement,
-        quarterly: quarterly,
-      },
-    });
-    console.log('Fetch response:', response); // Log the response
-
-    if (!response.data) {
-      const message = `An error has occurred: ${response.status}`;
-      throw new Error(message);
-    }
-    return response.data;
-  } catch (error) {
-    if (error.response) {
-      // Server responded with a status other than 200 range
-      console.error("Error fetching financial statement:", error.response.status, error.response.data);
-      throw new Error(`Error fetching financial statement: ${error.response.status}`);
-    } else if (error.request) {
-      // Request was made but no response received
-      console.error("Error fetching financial statement:", error.request);
-      throw new Error("Error fetching financial statement: No response received");
-    } else {
-      // Something else caused an error
-      console.error("Error fetching financial statement:", error.message);
-      throw new Error(`Error fetching financial statement: ${error.message}`);
-    }
-  }
-}
-
-export function transposeData(data) {
-  const transposed = {};
-  data.forEach((row, rowIndex) => {
-    Object.entries(row).forEach(([key, value]) => {
-      const formattedKey = key.replace(/_/g, ' ').replace(/\b\w/g, char => char.toUpperCase());
-      if (!transposed[formattedKey]) {
-        transposed[formattedKey] = [];
-      }
-      transposed[formattedKey][rowIndex] = value;
-    });
+  const response = await axios.get(`${pyBackEnd}/financials`, {
+    params: {
+      symbol: stockSymbol,
+      statement: statement,
+      quarterly: quarterly,
+    },
   });
-
-  return Object.entries(transposed).map(([key, values]) => ({ metric: key, ...values }));
-}
-
-export function formatFinancialData(data) {
-  let formattedData = "Financial Statement Data:\n";
-  data.forEach(row => {
-    const metric = row.metric;
-    const values = Object.values(row).slice(1); // exclude the 'metric' key
-    formattedData += `${metric}: ${values.join(', ')}\n`;
-  });
-  return formattedData;
+  return response.data;
 }
 
 export async function getFamaFrenchData(startDate, endDate, stockWeights) {
@@ -157,7 +114,7 @@ export async function getFamaFrenchData(startDate, endDate, stockWeights) {
   return response.data;
 }
 
-export function idbPromise(storeName, method, object) {
+export function idbPromise(stockWeights, method, object) {
   return new Promise((resolve, reject) => {
     const request = window.indexedDB.open("stockWeights", 1);
 
@@ -170,47 +127,35 @@ export function idbPromise(storeName, method, object) {
 
     request.onerror = function (e) {
       console.log("There was an error");
-      reject("Error in IDB request");
     };
 
     request.onsuccess = function (e) {
       db = request.result;
-      tx = db.transaction(storeName, "readwrite");
-      store = tx.objectStore(storeName);
+      tx = db.transaction("stockWeights", "readwrite");
+      store = tx.objectStore("stockWeights");
 
       db.onerror = function (e) {
         console.log("error", e);
-        reject("Database error");
       };
-
-      switch (method) {
-        case "put":
-          store.put(object);
-          resolve(object);
-          break;
-        case "get":
-          const all = store.getAll();
-          all.onsuccess = function () {
-            resolve(all.result);
-          };
-          break;
-        case "delete":
-          console.log("delete object", object.toString());
-          store.delete(object.toString());
-          resolve(object);
-          break;
-        default:
-          reject("Invalid method");
-          break;
+      if (method === "put") {
+        store.put(object);
       }
-
+      if (method === "get") {
+        const all = store.getAll();
+        all.onsuccess = function () {
+          resolve(all.result);
+        };
+      }
+      if (method === "delete") {
+        console.log("delete object", object.toString());
+        store.delete(object.toString());
+      }
       tx.oncomplete = function () {
         db.close();
       };
     };
   });
 }
-
 export async function getStockObject(
   userData,
   stockData,
@@ -236,14 +181,17 @@ export async function getStockObject(
         ...stockWeights,
         portfolio_id: userData?.user?.id,
       });
+  
 
       dispatch({
         type: SET_STOCK_WEIGHTS,
         payload: stockWeights,
       });
+      
     } catch (err) {
       console.log(err);
     }
   }
   setStockWeights(stockWeights);
+
 }
